% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logoplot_functions.R
\name{consensusMatrix_function}
\alias{consensusMatrix_function}
\title{Wrapper around consensusMatrix()}
\usage{
consensusMatrix_function(file, right_align = FALSE)
}
\arguments{
\item{file}{A fastq file}

\item{right_align}{Logical statement to 'right justify' all sequences found in the fastq file.  defaults to FALSE}
}
\value{
Returns a matrix of DNA base frequencies
}
\description{
Wrapper around consensusMatrix() from the ShortRead package to import sequences from a fastq file and (optionally) align them at the 5' (left) or 3' (right) DNA base
depends on packages Biostrings, data.table, ShortRead
}
\examples{
consensusMatrix_function('path/to/file.fastq', right_align = TRUE)
}
\author{
Justin Lin, \email{justin.jnn.lin@gmail.com}
}
\keyword{utilities}
