% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logoplot_functions.R
\name{make_logoplots}
\alias{make_logoplots}
\title{Creates a logoplot grob}
\usage{
make_logoplots(mat, method, label = "some_generic_sample", description)
}
\arguments{
\item{mat}{A matrix object from consensusMatrix_function()}

\item{method}{String value of 'bit' or 'prob' for ggseqlogo::geom_logo(), indicating plot method}

\item{label}{Plot title}

\item{description}{Plot description}
}
\value{
Returns a gtable object consisting of read logoplot and depth barplot
}
\description{
Creates a grob object (containing a read logoplot and depth barplot) from a consensusMatrix_function() matrix object
depends on ggplot2, ggpmisc, ggseqlogo
}
\examples{
make_logoplots(mat = mat, method = 'bits', label = 'tumor_sample', description = 'tumor_sample.fastq.gz')
}
\author{
Justin Lin, \email{justin.jnn.lin@gmail.com}
}
\keyword{graphics}
