% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_grobs.R
\name{print_grobs}
\alias{print_grobs}
\title{Print a pdf from a list of grobs,}
\usage{
print_grobs(grobs, output_filename, plot_width = 8, plot_height = 3)
}
\arguments{
\item{grobs}{A list of grobs}

\item{output_filename}{Name of the output file, including full path}

\item{plot_width}{Width (in inches) to draw each plot. Defaults to 8.}

\item{plot_height}{Height (in inches) to draw each plot. Defaults to 3.}
}
\description{
Creates a pdf from a list of grobs e.g. the output of make_logoplots().
}
\examples{
print_grobs(grobs, '/path/to/output.pdf', plot_width = 8, plot_height = 3)
}
\author{
Justin Lin, \email{justin.jnn.lin@gmail.com}
}
\keyword{graphics}
